requirements.txt
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
shap==0.49.1
python-dotenv==1.0.0
joblib==1.3.2
ipywidgets==8.1.5
python-dateutil==2.8.2
PyYAML==6.0.1

# APIs
openmeteo-requests==1.1.0
requests-cache==1.1.1
retry-requests==2.0.0
requests==2.31.0
openmeteo_sdk==1.23.0

# Feature Store 
hopsworks==4.2.10

# Dashboard & API
streamlit==1.29.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0